\name{convotron-package}
\alias{convotron-package}
\alias{convotron}
\docType{package}
\title{
Some convolutions in native code (C++)
}
\description{
This is demonstration software for native code implementations of convolution routines in the time and frequency domains.

}
\details{
\tabular{ll}{
Package: \tab convotron\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-10-17\cr
License: \tab GPL (>= 2)\cr
}
~~ An overview of how to use the package, including the most important ~~
TBD
~~ functions ~~
}
\author{
Daniel Pressel

Maintainer: Daniel Pressel <dpressel@gmail.com>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ convotron }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%% ~~ simple examples of the most important functions ~~
}
